Rename top-level package "subcommands" to "rpc"
