+++ cmd/plakar/subcommands/backup/backup.go
@@ -25,8 +25,8 @@ import (
 	"github.com/PlakarKorp/plakar/appcontext"
 	"github.com/PlakarKorp/plakar/cmd/plakar/subcommands"
 	"github.com/PlakarKorp/plakar/repository"
-	api_subcommands "github.com/PlakarKorp/plakar/subcommands"
-	"github.com/PlakarKorp/plakar/subcommands/backup"
+	"github.com/PlakarKorp/plakar/rpc"
+	"github.com/PlakarKorp/plakar/rpc/backup"
 	"github.com/gobwas/glob"
 )
 
@@ -45,7 +45,7 @@ func (e *excludeFlags) Set(value string) error {
 	return nil
 }
 
-func parse_cmd_backup(ctx *appcontext.AppContext, repo *repository.Repository, args []string) (api_subcommands.Subcommand, error) {
+func parse_cmd_backup(ctx *appcontext.AppContext, repo *repository.Repository, args []string) (rpc.RPC, error) {
 	var opt_tags string
 	var opt_excludes string
 	var opt_exclude excludeFlags
