+++ rpc/backup/stdio.go
