+++ rpc/cat/cat.go
