+++ rpc/info/snapshot.go
