+++ rpc/ls/ls.go
