+++ rpc/subcommands.go
@@ -1,11 +1,11 @@
-package subcommands
+package rpc
 
 import (
 	"github.com/PlakarKorp/plakar/appcontext"
 	"github.com/PlakarKorp/plakar/repository"
 )
 
-type Subcommand interface {
+type RPC interface {
 	Name() string
 	Execute(ctx *appcontext.AppContext, repo *repository.Repository) (int, error)
 }
